chore: split job-worker context.ts into multiple files
