+++ packages/job-worker/src/workers/context/util.ts
@@ -1,0 +1,1 @@
+export type QueueJobFunc = (queueName: string, jobName: string, jobData: unknown) => Promise<void>
