+++ packages/job-worker/src/workers/studio/child.ts
@@ -5,7 +5,8 @@ import { createMongoConnection, getMongoCollections, IDirectCollections } from '
 import { unprotectString } from '@sofie-automation/corelib/dist/protectedString'
 import { setupApmAgent, startTransaction } from '../../profiler'
 import { InvalidateWorkerDataCache, invalidateWorkerDataCache, loadWorkerDataCache, WorkerDataCache } from '../caches'
-import { QueueJobFunc, JobContextImpl } from '../context'
+import { JobContextImpl } from '../context/JobContextImpl'
+import { QueueJobFunc } from '../context/util'
 import { AnyLockEvent, LocksManager } from '../locks'
 import { FastTrackTimelineFunc, LogLineWithSourceFunc } from '../../main'
 import { interceptLogging, logger } from '../../logging'
